suse-commit: 41b905d2759cd8d27708a9cc7a88a9cd932e4878

+++ tools/objtool/elf.h
@@ -61,7 +61,7 @@ struct rela {
 	struct symbol *sym;
 	unsigned int type;
 	unsigned long offset;
-	int addend;
+	long addend;
 	bool jump_table_start;
 };
